// src/interfaces/portal/MintcapCreaturesIn.ts
export interface MintcapCreaturesIn {
    isForUser: null | boolean;
    type: null | string;
    tier: null | number | string;
    cohort: null | number;
    muscles: null | number;
    lungs: null | number;
    heart: null | number;
    belly: null | number;
    buttocks: null | number;
    isExpiredSoon: null | boolean;
    page?: number;
}